preparations for video downloading
