+++ Cargo.toml
@@ -40,6 +40,7 @@ tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
 walkdir = "2.3.3"
 webbrowser = "0.8.10"
 youtube_dl = { version = "0.8.1", features = ["tokio", "downloader"] }
+url = { version = "2.3.1", features = ["serde"] }
 
 [target.'cfg(unix)'.dependencies]
 tar = "0.4.38"
