+++ src/service/commands/mod.rs
@@ -1,0 +1,5 @@
+pub mod ffmpeg;
+mod ffprobe;
+mod yt_dlp;
+
+pub use self::ffprobe::ffprobe;
