+++ src/service/music/mod.rs
@@ -1,0 +1,5 @@
+mod beats;
+mod download;
+
+pub use self::beats::{detect_beats, parse_beats, Beats};
+pub use self::download::{MusicDownloadService, SongInfo};
